# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: oKidd Educational Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /schools:
    get:
      operationId: listSchools
      tags: [schools]
      responses:
        "200":
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      operationId: createSchool
      tags: [schools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{id}:
    get:
      operationId: getSchool
      tags: [schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: School
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
    put:
      operationId: updateSchool
      tags: [schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SchoolUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
  /schools/{id}/toggle-status:
    patch:
      operationId: toggleSchoolStatus
      tags: [schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"

  # Branches
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      parameters:
        - name: schoolId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Grade Levels
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    put:
      operationId: updateBook
      tags: [books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Lessons
...
  /content:
    get:
      operationId: listContent
      tags: [content]
      parameters:
        - name: lessonId
          in: query
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: bookId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of content
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Content"
    post:
      operationId: createContent
      tags: [content]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
  /content/{id}:
    put:
      operationId: updateContent
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Content"
    delete:
      operationId: deleteContent
      tags: [content]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
        schoolId:
          type: ["integer", "null"]
        nationalId:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
        schoolName:
          type: ["string", "null"]
...
    School:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: integer
        name:
          type: string
        userId:
          type: ["integer", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
        branchCount:
          type: integer
        studentCount:
          type: integer
        teacherCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    SchoolInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        userId:
          type: integer
        address:
          type: string
        phone:
          type: string
    SchoolUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        status:
          type: string
    Branch:
      type: object
      required: [id, schoolId, name]
      properties:
        id:
          type: integer
        schoolId:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        studentCount:
          type: integer
        createdAt:
          type: ["string", "null"]
    BranchInput:
      type: object
      required: [schoolId, name]
      properties:
        schoolId:
          type: integer
        name:
          type: string
        address:
          type: string
...
    Book:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
        lessonCount:
          type: integer
        monthlyFee:
          type: number
        gradeLevel:
          type: ["string", "null"]
        academicStage:
          type: ["string", "null"]
        isPreset:
          type: boolean
        createdAt:
          type: ["string", "null"]
    BookInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        lessonCount:
          type: integer
        monthlyFee:
          type: number
        gradeLevel:
          type: string
        academicStage:
          type: string
        isPreset:
          type: boolean
    BookUpdate:
      type: object
      properties:
        title:
          type: string
        lessonCount:
          type: integer
        monthlyFee:
          type: number
        gradeLevel:
          type: string
        academicStage:
          type: string
...
    Content:
      type: object
      required: [id, title, type]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        url:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        apiEndpoint:
          type: ["string", "null"]
        lessonId:
          type: ["integer", "null"]
        classId:
          type: ["integer", "null"]
        bookId:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: ["string", "null"]
    ContentInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
        url:
          type: string
        filePath:
          type: string
        apiEndpoint:
          type: string
        lessonId:
          type: integer
        classId:
          type: integer
        bookId:
          type: integer
        description:
          type: string
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
        url:
          type: string
        filePath:
          type: string
        description:
          type: string
        lessonId:
          type: ["integer", "null"]
        classId:
          type: ["integer", "null"]
        bookId:
          type: ["integer", "null"]
...