# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * oKidd Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  title: string;
  lessonCount?: number;
  monthlyFee?: number;
  gradeLevel?: string;
  academicStage?: string;
  isPreset?: boolean;
}