# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * oKidd Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  schoolId: number;
  name: string;
  /** @nullable */
  address?: string | null;
  studentCount?: number;
  /** @nullable */
  createdAt?: string | null;
}