# lib/api-zod/src/generated/types/schoolInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * oKidd Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SchoolInput {
  name: string;
  userId?: number;
  address?: string;
  phone?: string;
}