# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * oKidd Educational Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  phone?: string | null;
  status: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  nationalId?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  schoolName?: string | null;
}